MixedModelsMakie.jl API

Caterpillar Plots

MixedModelsMakie.RanefInfo — Type

RanefInfo

Information on random effects conditional modes/means, variances, etc.

Used for creating caterpillar plots.

Note

This functionality may be moved upstream into MixedModels.jl in the near future.

MixedModelsMakie.ranefinfo — Function

ranefinfo(m::LinearMixedModel)

Return a NamedTuple{fnames(m), NTuple(k, RanefInfo)} from model m

MixedModelsMakie.caterpillar — Function

caterpillar(m::LinearMixedModel, gf::Symbol)

Returns a Figure of a "caterpillar plot" of the random-effects means and prediction intervals

A "caterpillar plot" is a horizontal error-bar plot of conditional means and standard deviations of the random effects.

MixedModelsMakie.caterpillar! — Function

caterpillar!(f::Figure, r::RanefInfo; orderby=1)

Add Axes of a caterpillar plot from r to f.

The order of the levels on the vertical axes is increasing orderby column of r.ranef, usually the (Intercept) random effects.

using CairoMakie
CairoMakie.activate!(type = "svg")
using MixedModels
using MixedModelsMakie
sleepstudy = MixedModels.dataset(:sleepstudy)

fm1 = fit(MixedModel, @formula(reaction ~ 1 + days + (1 + days|subj)), sleepstudy)
subjre = ranefinfo(fm1)[:subj]

caterpillar!(Figure(; resolution=(800,600)), subjre)

caterpillar!(Figure(; resolution=(800,600)), subjre; orderby=2)

Shrinkage Plots

MixedModelsMakie.CoefByGroup — Type

CoefByGroup

Coefficients for groups of observations defined by levels of a grouping factor

Fields include:

cnames: Vector{String} of the column (or coefficient) names

levels: Vector of the levels of the grouping factor. Usually Strings but not necessarily.

fixed: AbstractVector{T} of the global OLS estimates for the coefficients in cnames only.

condmodes: AbstractMatrix{T} of size length(levels) x length(cnames) of the conditional means/modes for the random effects

grpest: similar to condmodes but allowing for Missing values, giving the within group OLS estimates of the coefficients.

Note

This functionality may be moved upstream into MixedModels.jl in the near future.

MixedModelsMakie.shrinkage — Function

shrinkage(m)

Return a NamedTuple{fnames(m), NTuple(k, CoefByGroup)} of the coefficients by group for the grouping factors

Missing docstring.

Missing docstring for shrinkageplot. Check Documenter's build log for details.

Missing docstring.

Missing docstring for shrinkageplot!. Check Documenter's build log for details.

using CairoMakie
using MixedModels
using MixedModelsMakie
sleepstudy = MixedModels.dataset(:sleepstudy)

fm1 = fit(MixedModel, @formula(reaction ~ 1 + days + (1 + days|subj)), sleepstudy)
shrink = shrinkage(fm1)[:subj]

shrinkageplot!(Figure(; resolution=(800,600)), shrink)